--- a/akd/agents/__init__.py
+++ b/akd/agents/__init__.py
@@ -1,3 +1,15 @@
-from ._base import BaseAgent, BaseAgentConfig, InstructorBaseAgent, LangBaseAgent
+from ._base import (
+    BaseAgent,
+    BaseAgentConfig,
+    InstructorBaseAgent,
+    LangBaseAgent,
+    LiteLLMInstructorBaseAgent,
+)
 
-__all__ = ["BaseAgent", "BaseAgentConfig", "InstructorBaseAgent", "LangBaseAgent"]
+__all__ = [
+    "BaseAgent",
+    "BaseAgentConfig",
+    "InstructorBaseAgent",
+    "LangBaseAgent",
+    "LiteLLMInstructorBaseAgent",
+]