--- a/akd/agents/risk/risk.py
+++ b/akd/agents/risk/risk.py
@@ -439,7 +439,13 @@ class RiskAgent(
 
         for risk_id in params.risk_ids:
             logger.info(f"Processing risk: {risk_id}")
-            self.memory.clear()
+
+            # start fresh if no tracking required
+            messages = [] if self.stateless else self.memory
+
+            # if empty, add system message
+            if not messages:
+                messages.append(self._default_system_message())
 
             # Combine risk definition and conversation into one user message
             risk_description = self._risk_map[risk_id]
@@ -449,15 +455,14 @@ class RiskAgent(
                 for i, (inp, outp) in enumerate(zip(params.inputs, params.outputs))
             )
 
-            user_prompt = f"""\
-            Risk ID: {risk_id}
-            Risk Description: {risk_description}
+            user_prompt = f"""
+Risk ID: {risk_id}
+Risk Description: {risk_description}\n
+Conversation:
+{conversation_text}
+"""
 
-            Conversation:
-            {conversation_text}
-            """
-
-            self.memory.append(
+            messages.append(
                 {
                     "role": "user",
                     "content": user_prompt,
@@ -467,7 +472,20 @@ class RiskAgent(
             # Use the per-risk schema here instead of the full Output Schema for the Risk Agent
             response = await self.get_response_async(
                 response_model=RiskCriteriaOutputSchema,
+                messages=messages,
             )
+
+            messages.append(
+                dict(
+                    role="assistant",
+                    content=response.model_dump_json(exclude={"type"}),
+                ),
+            )
+
+            # update memory only if stateful
+            if not self.stateless:
+                self._memory = messages
+
             logger.info(f"Judge criteria obtained for risk: {risk_id}")
             criteria_by_risk[risk_id] = response.criteria