--- a/akd/agents/search/components/instruction_builder.py
+++ b/akd/agents/search/components/instruction_builder.py
@@ -8,7 +8,7 @@ from loguru import logger
 from pydantic import Field
 
 from akd._base import InputSchema, OutputSchema
-from akd.agents._base import BaseAgentConfig, InstructorBaseAgent
+from akd.agents._base import BaseAgentConfig, LiteLLMInstructorBaseAgent
 from akd.configs.prompts import RESEARCH_INSTRUCTION_AGENT_PROMPT
 
 
@@ -17,10 +17,12 @@ class InstructionBuilderInputSchema(InputSchema):
 
     query: str = Field(..., description="Research query to build instructions for")
     context: Optional[str] = Field(
-        default=None, description="Additional context for instruction building"
+        default=None,
+        description="Additional context for instruction building",
     )
     clarifications: Optional[List[str]] = Field(
-        default=None, description="Clarifications gathered from the user/loop"
+        default=None,
+        description="Clarifications gathered from the user/loop",
     )
 
 
@@ -28,11 +30,13 @@ class InstructionBuilderOutputSchema(OutputSchema):
     """Output schema for instruction builder agent."""
 
     research_instructions: str = Field(
-        ..., description="Detailed research instructions"
+        ...,
+        description="Detailed research instructions",
     )
     search_strategy: str = Field(..., description="Recommended search strategy")
     key_concepts: List[str] = Field(
-        default_factory=list, description="Key concepts to focus on"
+        default_factory=list,
+        description="Key concepts to focus on",
     )
 
 
@@ -61,14 +65,17 @@ class InstructionBuilderComponent:
         self.debug = debug
 
         # Create internal instructor agent for instruction building
-        self._agent = InstructorBaseAgent[
-            InstructionBuilderInputSchema, InstructionBuilderOutputSchema
+        self._agent = LiteLLMInstructorBaseAgent[
+            InstructionBuilderInputSchema,
+            InstructionBuilderOutputSchema,
         ](config=self.config, debug=debug)
         self._agent.input_schema = InstructionBuilderInputSchema
         self._agent.output_schema = InstructionBuilderOutputSchema
 
     async def process(
-        self, query: str, clarifications: Optional[List[str]] = None
+        self,
+        query: str,
+        clarifications: Optional[List[str]] = None,
     ) -> str:
         """
         Build detailed research instructions from query and clarifications.
@@ -93,18 +100,18 @@ class InstructionBuilderComponent:
             preview_context = (instruction_input.context or "")[:200]
             preview_clar = ("; ".join(clarifications or []))[:200]
             logger.debug(
-                f"InstructionBuilder input preview | query: {query[:200]} | context: {preview_context} | clarifications: {preview_clar}"
+                f"InstructionBuilder input preview | query: {query[:200]} | context: {preview_context} | clarifications: {preview_clar}",
             )
 
         instruction_output = await self._agent.arun(instruction_input)
 
         if self.debug:
             logger.debug(
-                f"Generated research instructions ({len(instruction_output.research_instructions)} chars)"
+                f"Generated research instructions ({len(instruction_output.research_instructions)} chars)",
             )
             logger.debug(f"Key concepts: {instruction_output.key_concepts}")
             logger.debug(
-                f"InstructionBuilder output preview | research_instructions: {instruction_output.research_instructions[:200]} | search_strategy: {instruction_output.search_strategy[:200]}"
+                f"InstructionBuilder output preview | research_instructions: {instruction_output.research_instructions[:200]} | search_strategy: {instruction_output.search_strategy[:200]}",
             )
 
         return instruction_output.research_instructions