… akd/agents/search/components/research_synthesis.py
… tests/agents/risk/test_risk_agent.py