### Summary :memo:
This pull request introduces a new `LiteLLMInstructorBaseAgent` to serve as the standard base class for agents that use `litellm` for structured output. The majority of existing agents have been refactored to inherit from this new base class. This change centralizes the `litellm` API call logic, simplifies individual agent implementations, and fixes a critical bug where the API key was not being passed correctly. Additionally, the project's default `max_tokens` setting has been adjusted to a more sensible value.

### Details

1.  **New Base Agent:** Created `LiteLLMInstructorBaseAgent` in `akd/agents/_base.py` to encapsulate the logic for making `litellm.acompletion` calls.
2.  **Agent Refactoring:** Updated `IntentAgent`, `QueryAgent`, `RelevancyAgent`, `RiskAgent`, and several search component agents to subclass from the new `LiteLLMInstructorBaseAgent` instead of the more generic `InstructorBaseAgent`.
3.  **Configuration Update:** Modified the default `max_tokens` in `akd/configs/project.py` from `120,000,000` to `120,000` to prevent excessive token usage.
4.  **API Export:** Exposed the new `LiteLLMInstructorBaseAgent` in `akd/agents/__init__.py` for broader use within the package.

### Bugfixes :bug:
* Fixed a bug where the `api_key` was not being passed to the `litellm.acompletion` call within the base agent, which would cause authentication failures. This is now correctly handled in the new `LiteLLMInstructorBaseAgent`.

### Checks
- [x] Tested Changes (all tests passed successfully locally)
- [ ] Stakeholder Approval