@NISH1001 2025-09-24
On an initial test script to compare older agent vs new one, this is the report:
[test_new_old_agent.py](https://github.com/user-attachments/files/22521779/test_new_old_agent.py)


```
🧪 AGENT COMPARISON TEST REPORT
==================================================
Model: gpt-4o-mini
Iterations per test: 1
Total comparisons: 10

📋 QueryAgent
------------------------------
  📊 Average Character Similarity: 0.948
  🧠 Average Semantic Similarity: 1.000
  ⚡ Average Performance Ratio: 0.775x
  🔄 Average Consistency Score: 0.948
  ✅ Success Rate (Old): 100.0%
  ✅ Success Rate (New): 100.0%

    🧪 basic_query
       Old: ✅ (0.966s)
       New: ✅ (1.234s)
       Character Similarity: 1.000
       Semantic Similarity: 1.000

    🧪 complex_query
       Old: ✅ (3.413s)
       New: ✅ (1.459s)
       Character Similarity: 0.845
       Semantic Similarity: 1.000
       Key Differences:
         • Field 'queries' differs: '['latest developments in quantum computing applications for drug discovery', 'quantum computing in m...' vs '['latest developments in quantum computing for drug discovery', 'quantum computing applications in m...'

    🧪 single_query
       Old: ✅ (0.842s)
       New: ✅ (0.523s)
       Character Similarity: 1.000
       Semantic Similarity: 1.000

📋 FollowUpQueryAgent
------------------------------
  📊 Average Character Similarity: 0.598
  🧠 Average Semantic Similarity: 0.964
  ⚡ Average Performance Ratio: 0.934x
  🔄 Average Consistency Score: 0.598
  ✅ Success Rate (Old): 100.0%
  ✅ Success Rate (New): 100.0%

    🧪 renewable_energy_followup
       Old: ✅ (2.897s)
       New: ✅ (2.796s)
       Character Similarity: 0.847
       Semantic Similarity: 1.000
       Key Differences:
         • Field 'reasoning' differs: 'These follow-up queries aim to explore deeper into the advancements in energy storage, compare solar...' vs 'These follow-up queries aim to explore deeper into the advancements in energy storage, compare solar...'

    🧪 research_gap_identification
       Old: ✅ (3.466s)
       New: ✅ (2.871s)
       Character Similarity: 0.207
       Semantic Similarity: 0.893
       Key Differences:
         • Field 'followup_queries' differs: '['What are the latest advancements in machine learning applications?', 'How do neural networks impro...' vs '['What are the latest advancements in machine learning applications?', 'How do neural networks compa...'
         • Field 'reasoning' differs: 'These follow-up queries aim to explore recent developments in machine learning, delve deeper into th...' vs 'These follow-up queries aim to explore recent developments in machine learning and neural networks, ...'
         • Field 'original_query_gaps' differs: '['Recent advancements in machine learning applications', 'Interpretability techniques in neural netw...' vs '['Recent advancements in machine learning applications', 'Comparison between neural networks and tra...'

    🧪 broad_topic_refinement
       Old: ✅ (2.185s)
       New: ✅ (2.205s)
       Character Similarity: 0.741
       Semantic Similarity: 0.998
       Key Differences:
         • Field 'reasoning' differs: 'These follow-up queries aim to explore specific adaptation strategies and their effectiveness, as we...' vs 'These follow-up queries aim to explore specific adaptation strategies in detail and understand the b...'

📋 MultiRubricRelevancyAgent
------------------------------
  📊 Average Character Similarity: 0.390
  🧠 Average Semantic Similarity: 0.981
  ⚡ Average Performance Ratio: 1.180x
  🔄 Average Consistency Score: 0.390
  ✅ Success Rate (Old): 100.0%
  ✅ Success Rate (New): 100.0%

    🧪 comprehensive_relevant
       Old: ✅ (5.182s)
       New: ✅ (4.298s)
       Character Similarity: 0.438
       Semantic Similarity: 0.976
       Key Differences:
         • Field 'reasoning_steps' differs: '['The query focuses on sustainable energy solutions, which aligns with the content discussing renewa...' vs '['The query focuses on sustainable energy solutions, which aligns with the content discussing renewa...'

    🧪 outdated_content
       Old: ✅ (6.358s)
       New: ✅ (9.726s)
       Character Similarity: 0.341
       Semantic Similarity: 0.986
       Key Differences:
         • Field 'reasoning_steps' differs: '["The topic of the query is 'artificial intelligence in healthcare', which is a broad and current fi...' vs '["The topic of the query is 'artificial intelligence in healthcare', which is a broad and current fi...'

📋 EstimationExtractionAgent
------------------------------
  📊 Average Character Similarity: 0.236
  🧠 Average Semantic Similarity: 0.991
  ⚡ Average Performance Ratio: 0.998x
  🔄 Average Consistency Score: 0.236
  ✅ Success Rate (Old): 100.0%
  ✅ Success Rate (New): 100.0%

    🧪 performance_metrics
       Old: ✅ (2.561s)
       New: ✅ (2.476s)
       Character Similarity: 0.267
       Semantic Similarity: 0.995
       Key Differences:
         • Field 'estimations' differs: '[{'answer': '22% efficiency under standard test conditions', 'related_knowledge': ['Estimated lifesp...' vs '[{'answer': '22% efficiency under standard test conditions', 'related_knowledge': ['Estimated lifesp...'

    🧪 research_findings
       Old: ✅ (2.772s)
       New: ✅ (2.851s)
       Character Similarity: 0.205
       Semantic Similarity: 0.988
       Key Differences:
         • Field 'estimations' differs: '[{'answer': '85% response rate with median progression-free survival of 18 months.', 'related_knowle...' vs '[{'answer': '85% response rate with median progression-free survival of 18 months.', 'related_knowle...'

🎯 OVERALL SUMMARY
--------------------
Average Character Similarity: 0.589
Average Semantic Similarity: 0.983
Average Performance Ratio: 0.948x
Average Consistency Score: 0.589

⚡ Performance: Similar between old and new agents
🟡 Outputs: Same meaning, different formatting (low character, high semantic)

📖 INTERPRETATION GUIDE
--------------------
Character Similarity: Exact JSON structure/content match
Semantic Similarity: Meaning and content similarity

Similarity Pattern Analysis:
  High Char + High Semantic = Identical responses
  Low Char + High Semantic = Same meaning, different format
  High Char + Low Semantic = Similar structure, different content
  Low Char + Low Semantic = Genuinely different responses
```